fix(catalog): stop truncating half credits to whole ones
`int(float(...))` on a parsed credit value, and `courses.credits` an integer
column. 129 of the 749 ingested courses are worth 1.5, so each one landed as 1.

The direction of the error is what matters. Entrepreneurship's capstone is two
1.5-credit courses the bulletin counts as three; stored as 1+1 the programme
reconciles to 28 against a stated 30, and a degree audit tells a student they owe
credits they have already earned. Executive Masters in Marketing is built almost
entirely from halves. The sequence planner's per-term credit cap was computing
against the same wrong numbers.

The column is now double precision and the parser keeps the fraction. Halves are
exactly representable in binary floating point and halves are the only fraction
this catalogue uses, so there is no rounding behaviour to reason about — worth
saying because "float for a quantity that must add up" is otherwise a fair thing
to object to.

Found while encoding the remaining programmes: the transcription of
Entrepreneurship would not reconcile, and the cause was the catalogue rather than
the transcription. Encoding on top of it would have produced twenty programmes
that each fail validation for a reason unrelated to what validation is for.

Verified: all three encoded programmes still reconcile to their stated totals,
pytest 361 with two new assertions over the real snapshot — one property-level,
one naming a specific pair so the failure is legible.

+++ api/app/intake/types.py
@@ -46,7 +46,7 @@ class ExtractedRow:
     # Everything below is best-effort; each may be absent without invalidating the row.
     title: str | None = None
     term: str | None = None
-    credits: int | None = None
+    credits: float | None = None
     grade: str | None = None
     # What the reader believes the student's relationship to this course is. Never derived
     # from a missing grade alone without saying so in `reasons`.
